Added functionality to main, useState handling, and LandingPage, Changed file names to better reflect uses
