+++ src/main.tsx
@@ -1,11 +1,21 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 
-import App from "./App.tsx";
+import AppRoot from "./App.tsx";
 import "./index.css";
 
-const root = document.getElementById("root");
-if (root !== null) {
+const rootElement = document.getElementById("root");
+if (rootElement) {
+  ReactDOM.createRoot(rootElement).render(
+    <React.StrictMode>
+      <AppRoot /> {/* Updated component reference */}
+    </React.StrictMode>,
+  );
+} else {
+  console.error("root element is missing");
+}
+
+/*if (root !== null) {
   ReactDOM.createRoot(root).render(
     <React.StrictMode>
       <App />
@@ -13,4 +23,4 @@ if (root !== null) {
   );
 } else {
   console.error("root element is missing");
-}
+}*/
